BUG FIXES: Fixed better IType guessing since NHibernateUtil::GuessType isn't that reliable. This fixed some issues with type mismatches in projections and restrictions (mostly with conditional projections).
POTENTIALLY BREAKING CHANGES:
Modified the onClause parameters for joins to not specify TSource as input, but rather TAlias, since this seems to be the way NHibernate interprets it anyway. In order to use TSource in the onClause for joins you seem to have to specify a root alias for TSource on the criteria level. I don't know if this is a bug in NHibernate or the intentional design.

Made it possible to aggregate on nullable types into non-nullable types.