Added a test for a real scenario.
Will probably remove the Aggregate.FromExpression method and let users implement it themselves or come up with a better solution for it (an over-ridable one). In order for the user code to be testable it must actually return a value which doesn't comply with the design principles for FlowQuery (Can't implement real logic/code for every DB feature supported).